+++ packages/client/test/unit/OrderMessages.test.ts
@@ -3,8 +3,8 @@ import { randomEthereumAddress } from '@streamr/test-utils'
 import { EthereumAddress, collect } from '@streamr/utils'
 import range from 'lodash/range'
 import without from 'lodash/without'
-import { OrderMessages } from '../../src/subscribe/OrderMessages'
 import { ResendOptions, Resends } from '../../src/subscribe/Resends'
+import { OrderMessages } from '../../src/subscribe/ordering/OrderMessages'
 import { fromArray } from '../../src/utils/GeneratorUtils'
 import { PushPipeline } from '../../src/utils/PushPipeline'
